Adjusted Build to include files to npm package
- added some barrels to simplify imports from package
