+++ .travis.yml
@@ -8,7 +8,7 @@ addons:
 cache: npm
 install:
   - npm install --ignore-scripts # Ignore Botmation CLI (no need for User walkthrough)
-  - npm run createbaseconfigfile # but, create the config file, was defaults
+  - npm run createbaseconfigfile # but, create the config file, with defaults
   - npm rebuild puppeteer # and run the scripts for puppeteer
 script:
   - jest
